backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
asyncpg==0.29.0
pydantic==2.5.0
python-dotenv==1.0.0
fastapi-cors==0.0.6
python-multipart==0.0.6